AUI-1177 Date disappears from input textbox when choosing the minimumDate from the DatePicker
This bug was happening because when the user clicks on the test field the datepicker code shows the calendar again and sets its dates to the ones in the field's value. Since the text field doesn't have the time information of the selected date (the default mask has only the day, month and year), the date object created from it would have a different time than the original (in the default case it would be the same as the time on the user's computer), though it would still be the same date. The calendar module considers all of its dates as having the time set to 12:00:00, so if the selected date had been equal to the minimimDate option, it would reject it when the user clicked the text field before 12:00:00 (similarly, it would reject dates equal to the maximumDate when the user clicked after 12:00:00).
So what I did was fix the dates being passed to the calendar's selectDates function, adjusting the timestamp to make sure that it won't be rejected when the date is equal to minimumDate or maximumDate.