AUI-1208 Cloned fields displaying hidden attributes
Cloned fields were only receiving the field properties of the original field during creation. Attributes like hiddenAttributes weren't being passed, so this configuration wouldn't be respected for cloned fields.
While fixing this for hiddenAttribute I noticed that this was already happening for all other non-property attributes. I checked these other attributes one by one and this fix is now making sure that the ones that need to be passed over to the cloned field are doing so.