AUI-1294 Make aui-colorpicker responsive
We're changing the structure of the HSV and HSVA colorpickers to use bootstrap grid classes instead of absolute positions.

Also enabling the user to specify the width of these colorpickers through the width attribute, in which case we won't have the default behavior of making it vertical when the screen is too small, it will always have the width specified by the user.

Finally, I took the chance to change the HSVA and RGB input fields to be of type 'number'.