AUI-1316 Can't pause carousel through button if pauseOnHover is true
The carousel's pause button wasn't working well when pauseOnHover was set to true and the menu was inside the carousel (which is the default configuration).

This is because hovering on the pause button would also hover on the carousel, pausing the slideshow. Because of this the user wasn't able to manually pause by clicking the button.

This fixes the issue by listening to mousenter and mouseleave events for the menu as well. If the mouse enters the menu after leaving the carousel's bounding box, that means that the menu is inside the bounding box itself. In this case we need to resume the slideshow so the user can use the buttons as expected. When leaving the menu to hover on top of the image again this pauses the slideshow as expected too.

There's another bug that was fixed here as well, which happens if the carousel is paused before the user hovers on it. In that case, when the user hovers out the carousel will be resumed, even though it was already paused before the hover action happened. I added a variable now that will make sure to only reume the slideshow on mouseleave if it was paused due to a mouseenter.