AUI-1318 Fixing test on IE8
One of the test cases for the image viewer was failing on IE8 when run together with other tests due to image caching. We need the image to not be on cache when this test case runs, so I've changed the image being used to make sure to one that wasn't being used in other tests.