AUI-1310 Fixing tooltip tests
Some tooltip tests were failing when run in Jenkins because of the viewport size used by the browsers being tested there. These tests need the buttons used in them to be inside the viewport, and one of them needs a button to be outside it. I've changed the layout of the index.html file to make sure that this would happen for these buttons, so now the tests pass there as well.