AUI-1442 First click on each toolbar button is ignored on touch devices
The toolbar only creates Button instances for its buttons on the first user interaction with them. On regular computers this happens when first mousing over a button or when focusing it through the keyboard. These actions can't be done on touch devices though, so the first interaction is already the click itself. Because of this the first click on the button doesn't trigger the button's click event on touch devices, since it hadn't been initialized yet.

To fix this I'm now always creating the button instance for touch devices instead of lazily creating them.