AUI-1300 Make aui-datatable work on touch devices
Making a few changes to datatable so it will work well on touch devices:
(1) Property lists used to have double click as the default edit event. I changed this so that it will use click instead when on a touch device.
(2) The datatable editors are now constrained to the viewport, to make sure they'll always show up in the small screens.

Also added datatable-scroll to the demo as an example of how to add scroll to the datatable.