AUI-1303 Fixing form builder interactions on touch enabled desktops
Desktop computers with touch screens couldn't click on a form builder field to open its settings anymore. That's because we removed this feature for touch enabled devices, to avoid problems with scrolling. That should not happen for desktop computers though, since they have larger screens and the mouse.
This is changing the touchEnabled checks on form builder to check for the mobile flag as well, like Robert Frampton suggested.