AUI-1593 Removing ie8 hack on ImageViewerBase
It turns out that the problem mentioned on ImageViewerBase wasn't cause by YUI itself. The reason why setting attributes to the rendered image didn't work on ie8 is because of the way the template was defined, with a closing "</img>" tag in the end. On ie8 that gets turned into a fragment instead of a single node, so setting attributes didn't work. I changed the template to close the img tag directly and it works now without the hack.