AUI-1549 Error in widget-modality.js
Currently, if a modal is showed on a page where the scrollbar is not at the top, it will scroll the page to the top and show up there instead of at the place where it was triggered.

This is because we're using WidgetModality together with WidgetPositionAlign. WidgetModality focuses the modal when it becomes visible, but it does so before WidgetPositionAlign positions it on the viewport, so the window gets scrolled to the top and only then the modal is positioned.

This is fixing the issue by just changing the order of the modal augmentations, putting WidgetModality at the bottom, so it only focuses the modal when it's ready and correctly positioned on the viewport.