AUI-1610 Changing predefinedValue for text fields doesn't update them
The wrapper that I added when starting to use bootstrap 3 has broken the updates of predefinedValue, as the code tries to set the value on the wrapper instead of the actual text field. We don't actually need the wrapper anymore though, as it's not a row anymore, so I'm just removing it here.