AUI-1612 Adding wrapper for textfield and textarea
When working on this bug I noticed that https://github.com/liferay/alloy-ui/commit/841a65ce71cf048140cc2696a7570626495ddd73 has broken the width feature for textfield and textarea. The wrapper was actually still necessary for that, bootstrap requires it. So I changed aui-form-builder-field to find the actual input inside a templateNode, and make changes on that instead. It now works well for both inputs with wrappers and without them.