AUI-1820 Fixes listener for key presses on aui-dropdown
The listener was being triggered for all key presses, not just `enter`, since it was listening to `keypress` instead of `key`.
Also, for triggers that were buttons or links the browser automatically triggers a `click` event when enter is pressed, and so `toggleContent` could be called twice (one for the `key` event and another for the `click` event. So to fix this I just added a call to `preventDefault`, so the `click` event won't be triggered after we've already handled the `key` one.