AUI-2074 Revert "AUI-2061 If there is no event, there is no need to a "See more" link."
This reverts commit 2e8b98056b50058cdc4d05dcb20d7eabb82fd704.

The new fix solves the problem without this commit. While this reverted change
is not wrong, it is difficult to know why the code was different in the first
place. It is better to not alter it, I guess. Anyway, an advantage of not
checking for the event is that the scheduler tends to fail very uglily; if it
checks, then errors can happen silently. That's what caused LPS-62778 for
example.